backend/requirements.txt
fastapi
uvicorn
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl
pytextrank
langdetect
requests
cohere
anthropic
python-dotenv